Make it easy to customize operators and columns.
Added a custom `operators()` and `columns()` method that can be overridden. This PR allows you to do this:

```
use philperusse\Filters\ColumnFilter as Filter;

class MyFilter extends Filter
{
    public function columns()
    {
        return [
            'name' => 'Name',
            'age' => 'Age',
        ];
    }

    public function operators()
    {
        return [
            '=' => '&equals;',
            '>' => '&gt;',
            '<' => '&lt;',
        ];
    }
    
}
```

You could take it a step further by allowing the `operators()` function to return an array like `['=', '>']` so we don't need to translate them.